changed code structure for better representation
