--- src/main/java/com/example/todo_app/hellow.java
@@ -1,12 +1,0 @@
-package com.example.todo_app;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class hellow {
-    @GetMapping(path = "/hellow")
-    public String say_hellow(){
-        return "hellow!";
-    }
-}
